Move pageinfo into `src/article-api` (#54620)
