+++ package.json
@@ -64,7 +64,7 @@
     "playwright-test": "playwright test --config src/fixtures/playwright.config.ts --project=\"Google Chrome\"",
     "post-lints": "tsx src/content-linter/scripts/post-lints.js",
     "postinstall": "cp package-lock.json .installed.package-lock.json && echo \"Updated .installed.package-lock.json\" # see husky/post-checkout and husky/post-merge",
-    "precompute-pageinfo": "tsx src/pageinfo/scripts/precompute-pageinfo.ts",
+    "precompute-pageinfo": "tsx src/article-api/scripts/precompute-pageinfo.ts",
     "prepare": "husky src/workflows/husky",
     "prettier": "prettier -w \"**/*.{ts,tsx,js,mjs,scss,yml,yaml}\"",
     "prettier-check": "prettier -c \"**/*.{ts,tsx,js,mjs,scss,yml,yaml}\"",
