+++ src/frame/middleware/api.ts
@@ -5,7 +5,7 @@ import events from '@/events/middleware.js'
 import anchorRedirect from '@/rest/api/anchor-redirect.js'
 import aiSearch from '@/search/middleware/ai-search'
 import search from '@/search/middleware/search-routes.js'
-import pageInfo from '@/pageinfo/middleware'
+import pageInfo from '#src/article-api/middleware/pageinfo.ts'
 import pageList from '@/article-api/middleware/pagelist'
 import webhooks from '@/webhooks/middleware/webhooks.js'
 import { ExtendedRequest } from '@/types'
